novo exercicio adicionado - Texas Holdem

+++ lista02/exe02Classes/fazMedia.cs
@@ -35,7 +35,7 @@ namespace exe02Classes
             return arrValorInt [i];
         }
 
-        public double getMediaPonderada ()
+        public float getMediaPonderada ()
         {
             return mediaPonderada;
         }
